Fixed journal to support writing multiple versions.  Working on transactional isolation.
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/trunk@18 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/java/com/bigdata/journal/DataDeletedException.java
@@ -13,9 +13,9 @@ public class DataDeletedException extends RuntimeException {
     /**
      * The persistent identifier.
      */
-    public final long id;
+    public final int id;
     
-    public DataDeletedException(long id) {
+    public DataDeletedException(int id) {
        
         super( "Data is deleted: id="+id );
